// src/react/redux/types/ui/dashboardVisualization.ts
export const DataVisualizationsList = [TableKey, TimeSeriesKey, HistogramKey, Histogram3DKey] as const;
export type DataVisTypes = (typeof DataVisualizationsList)[number];
// ...
type VisDataSetType<T extends DataVisTypes> =
  | StatelessVisDataSetType<T>
  | {
      type: T;
      state: "Loading";
    };

type VisPropertiesType<T extends VisTypes> = z.infer<(typeof VisPropertiesMap)[T]>;

type VisDataRefreshParamsType<T extends DataVisTypes> = VisDataRefreshParamsMap[T];

type VisDataSetZodType<T> = T extends DataVisTypes ? ZodType<VisDataSetType<T>, ZodTypeDef, VisDataSetType<T>> : never;

export type DataSet<T extends DataVisTypes> = VisDataSetType<T>;

export type VisualizationDataContextError = "API_ERROR" | "NO_DATA_FOR_DATA_SET"; //appLabels.VisualizationErrorMessages.dataNotAvailable

type VisualizationDataContextPart = {
  datasets: DataSetArray | null;
  dashboardId: string;
  visualizationId: string;
  hasError: boolean;
  connectionIsEmpty: boolean;
  errorCode: VisualizationDataContextError | null;
};
// ...
export type OnDataSetUpdatedCallback = ({ message }: { message: SignalRDataSetUpdatedEvent }) => void;

export type VisualizationDataContextState<T extends DataVisTypes> = VisualizationDataContextPart & {
  schemaStore: DataSetSchemaStore<T> | null;
  onDataSetUpdatedCallback: OnDataSetUpdatedCallback | null;
  hasConnection: boolean;
  properties: DataVisualizationProperties<T>;
};

export type VisualizationDataContextData<T extends DataVisTypes> = VisualizationDataContextPart & {
  setOnDataSetUpdatedCallback: (onDataSetUpdatedCallback: OnDataSetUpdatedCallback | null) => void;
  getSchema: (dataSetId: DataSetId) => Promise<DataSet<T>>;
  getDataRange: (dataSetId: string) => Promise<DataRange>;
};

export type CreateVisualizationParams = {
  payload: CreateVisualizationPayload;
  callback?: (visualizationId: string) => void;
};

export type DataSetId = { dataSetId: string };