// src/react/components/dashboard-visualization/components/TimeSeries2/types.ts
export type LineChartData = {
  data: LineChartDataSeries;
  label: string;
  min: number;
  max: number;
  dataSetId: string;
};

export type RequestAndPrepareColumnData = (
  params: CommonParams<"TimeSeries"> & {
    dataRange: DataRange | null;
    pointCount: number;
  }
) => Promise<LineChartData[] | null>;

export type LineChartDataWithColor = LineChartData & { color: string };
// ...
export type HiddenColumnsKey = `${string}-${string}`;

export type HiddenColumnsMap = Record<HiddenColumnsKey, boolean>;

export type StoredChartState = {
  range: Range | null;
  showMarkers: boolean;
  hiddenColumns: HiddenColumnsMap;
  tooltipsEnabled: boolean;
  showMinimap: boolean;
};

export type TimeSeriesContextData = {
  chartData: ChartData;
  refreshActionsRef: RefObject<(() => void) | null>;
  visError: VIS_ERROR_TYPES | null;
  visMessage: string | null;
  ref: RefObject<HTMLDivElement | null>;
  minimapRef: RefObject<HTMLDivElement | null>;
  legendRef: RefObject<HTMLDivElement | null>;
};

export type GetColumnData<T extends DataVisTypes> = (
  params: CommonParams<T> & {
    width: number;
    visibleRange: NumberRange | Range | null;
  }
) => Promise<
  | {
      state: "Loaded";
      data: {
        chartColumns: LineChartDataWithColor[];
        xLimits: Range;
        minimapColumns: LineChartDataWithColor[];
      };
    }
  | { state: "Loading" }
  | null
>;

export type ChartData = {
  resizeObserver: ResizeObserver | null;
  numtraces: number;
  xAxisLabel: string;
  yAxisLabel: string;
  range: Range | null;
  fullRange: Range;
  dataSetId: string | null;
  datasetLabels: string[];
  chartColumns: LineChartDataWithColor[];
  minimapChartColumns: LineChartDataWithColor[];
  chartSurface: SciChartSurface | null;
  wasmContext: TSciChart | null;
  overview: SciChartOverview | null;
  xySeriesArray: XyDataSeries[];
  minimapXySeriesArray: XyDataSeries[];
};